requirements.txt
streamlit>=1.28.0,<1.29.0
boto3>=1.29.2
python-dotenv==1.0.0
Pillow==10.4.0
urllib3==2.5.0
cfn-lint==1.40.3